Added Beamdyn specific case

+++ OpenFast/f90wrap_test_tools.f90
@@ -4,9 +4,9 @@ subroutine f90wrap_calcwmparameters(params, angle, n)
     use test_tools, only: calcwmparameters
     implicit none
     
-    real(4), dimension(3) :: params
-    real(4) :: angle
-    real(4), dimension(3) :: n
+    real(8), dimension(3) :: params
+    real(8) :: angle
+    real(8), dimension(3) :: n
     call calcwmparameters(params=params, angle=angle, n=n)
 end subroutine f90wrap_calcwmparameters
 
@@ -14,9 +14,9 @@ subroutine f90wrap_calcrotationmatrix(angle, ret_calcrotationmatrix, axis)
     use test_tools, only: calcrotationmatrix
     implicit none
     
-    real(4), intent(in) :: angle
-    real(4), dimension(3,3), intent(out) :: ret_calcrotationmatrix
-    real(4), intent(in), dimension(3) :: axis
+    real(8), intent(in) :: angle
+    real(8), dimension(3,3), intent(out) :: ret_calcrotationmatrix
+    real(8), intent(in), dimension(3) :: axis
     ret_calcrotationmatrix = calcrotationmatrix(angle=angle, axis=axis)
 end subroutine f90wrap_calcrotationmatrix
 
@@ -24,7 +24,7 @@ subroutine f90wrap_identity(ret_identity)
     use test_tools, only: identity
     implicit none
     
-    real(4), dimension(3,3), intent(out) :: ret_identity
+    real(8), dimension(3,3), intent(out) :: ret_identity
     ret_identity = identity()
 end subroutine f90wrap_identity
 
@@ -32,8 +32,8 @@ subroutine f90wrap_ronxaxis(ret_ronxaxis, angle)
     use test_tools, only: ronxaxis
     implicit none
     
-    real(4), dimension(3,3), intent(out) :: ret_ronxaxis
-    real(4) :: angle
+    real(8), dimension(3,3), intent(out) :: ret_ronxaxis
+    real(8) :: angle
     ret_ronxaxis = ronxaxis(angle=angle)
 end subroutine f90wrap_ronxaxis
 
@@ -41,7 +41,7 @@ subroutine f90wrap_getmassmatrix(ret_getmassmatrix)
     use test_tools, only: getmassmatrix
     implicit none
     
-    real(4), dimension(6,6), intent(out) :: ret_getmassmatrix
+    real(8), dimension(6,6), intent(out) :: ret_getmassmatrix
     ret_getmassmatrix = getmassmatrix()
 end subroutine f90wrap_getmassmatrix
 
@@ -49,7 +49,7 @@ subroutine f90wrap_getstiffnessmatrix(ret_getstiffnessmatrix)
     use test_tools, only: getstiffnessmatrix
     implicit none
     
-    real(4), dimension(6,6), intent(out) :: ret_getstiffnessmatrix
+    real(8), dimension(6,6), intent(out) :: ret_getstiffnessmatrix
     ret_getstiffnessmatrix = getstiffnessmatrix()
 end subroutine f90wrap_getstiffnessmatrix
 
@@ -57,7 +57,7 @@ subroutine f90wrap_getgravityinz(ret_getgravityinz)
     use test_tools, only: getgravityinz
     implicit none
     
-    real(4), dimension(3), intent(out) :: ret_getgravityinz
+    real(8), dimension(3), intent(out) :: ret_getgravityinz
     ret_getgravityinz = getgravityinz()
 end subroutine f90wrap_getgravityinz
